requirements.txt
streamlit
faiss-cpu
torch
sentence-transformers
python-pptx
PyMuPDF
python-docx